hardware/requirements.txt
websockets==11.0.3
pyserial==3.5
asyncio
requests>=2.25.0
RPi.GPIO>=0.7.0; platform_machine=="armv7l"
